add firebase to next-client
