+++ .gitignore
@@ -43,7 +43,7 @@ coverage
 !.vscode/launch.json
 !.vscode/extensions.json
 
-# misc
+# express-server
 apps/express-server/public
 apps/express-server/uploads
