+++ apps/next-client/package.json
@@ -24,6 +24,7 @@
     "@sentry/nextjs": "^9",
     "@tanstack/react-table": "^8.21.3",
     "axios": "^1.8.4",
+    "firebase": "^11.6.0",
     "graphql": "^16.10.0",
     "graphql-hooks": "^8.2.0",
     "graphql-hooks-memcache": "^3.2.0",
