+++ apps/next-client/src/app/api/person/route.ts
@@ -1,6 +1,6 @@
 import { type NextRequest, NextResponse } from 'next/server';
 import { PersonModel } from '@csl/mongo-models';
-import mongoDB from '@/mongoDB';
+import { mongoDBService } from '@/services';
 
 export async function GET(request: NextRequest) {
   /**
@@ -23,7 +23,7 @@ export async function GET(request: NextRequest) {
   const recordsPerPage = limit ? Number(limit) : 10;
 
   try {
-    await mongoDB.connect();
+    await mongoDBService.connect();
     const allData = await PersonModel.find()
       .skip((pageNum - 1) * recordsPerPage)
       .limit(recordsPerPage);
