+++ apps/next-client/src/constants/firebase.config.ts
@@ -1,0 +1,9 @@
+export const firebaseConfig = {
+  apiKey: 'AIzaSyCvCUEk0CJLBZzT6VTokzSepuDnP_Fj2o0',
+  authDomain: 'client-server-libs.firebaseapp.com',
+  projectId: 'client-server-libs',
+  storageBucket: 'client-server-libs.firebasestorage.app',
+  messagingSenderId: '970682162267',
+  appId: '1:970682162267:web:68494fdfe9c308660e737e',
+  measurementId: 'G-LNY3SYRQ7H'
+};
