+++ apps/next-client/src/services/index.ts
@@ -1,0 +1,1 @@
+export { default as mongoDBService } from './mongoDB';
