+++ docs/Firebase.md
@@ -1,0 +1,5 @@
+# Firebase
+
+1.  You can add multiple **web**, **android** or **ios** apps for a single firebase project.
+
+2.  Generate the Admin SDK credentials for your project under `Project Settings > Service Accounts > Firebase Admin SDK`. These credentials must never be exposed.
